ENG-1001: Set up deployment test suite
It's somewhat "best effort". For example, we're currently relying on the configs being correct for some things (like council and revenue pool addresses). We can improve stuff like that as we improve the deployment processes.  (In the meantime, I guess we can manually update the configs to make tests green.)

We still need governance contract tests, and a GitHub action, and other stuff, but this is enough to start testing the concept.